+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_help.py
@@ -2371,6 +2371,11 @@ helps['network local-gateway update'] = """
             az network local-gateway update -g MyResourceGroup -n MyLocalGateway \\
                 --local-address-prefixes 10.0.0.0/24 20.0.0.0/24 30.0.0.0/24
 """
+
+helps['network local-gateway wait'] = """
+    type: command
+    short-summary: Place the CLI in a waiting state until a condition of the local gateway is met.
+"""
 # endregion
 
 # region Network Interface (NIC)
@@ -3038,7 +3043,6 @@ helps['network traffic-manager endpoint create'] = """
     short-summary: Create a traffic manager endpoint.
     parameters:
         - name: --geo-mapping
-          short-summary: Space-separated list of country/region codes mapped to this endpoint when using the 'Geographic' routing method.
           populator-commands:
           - az network traffic-manager endpoint show-geographic-hierarchy
     examples: