+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_params.py
@@ -27,7 +27,7 @@ from azure.cli.command_modules.network._validators import (
     get_network_watcher_from_vm, get_network_watcher_from_location,
     get_asg_validator, get_vnet_validator, validate_ip_tags, validate_ddos_name_or_id)
 from azure.mgmt.network.models import ApplicationGatewaySslProtocol
-from azure.mgmt.trafficmanager.models import MonitorProtocol
+from azure.mgmt.trafficmanager.models import MonitorProtocol, ProfileStatus
 from azure.cli.command_modules.network._completers import (
     subnet_completion_list, get_lb_subresource_completion_list, get_ag_subresource_completion_list,
     ag_url_map_rule_completion_list, tm_endpoint_completion_list)
@@ -341,9 +341,10 @@ def load_arguments(self, _):
         for item in ['record_type', 'record_set_type']:
             c.argument(item, ignore_type, validator=validate_dns_record_type)
 
-    with self.argument_context('network dns record-set create') as c:
-        c.argument('ttl', help='Record set TTL (time-to-live)')
-        c.argument('if_none_match', help='Create the record set only if it does not already exist.', action='store_true')
+    for item in ['', 'a', 'aaaa', 'caa', 'cname', 'mx', 'ns', 'ptr', 'srv', 'txt']:
+        with self.argument_context('network dns record-set {} create'.format(item)) as c:
+            c.argument('ttl', help='Record set TTL (time-to-live)')
+            c.argument('if_none_match', help='Create the record set only if it does not already exist.', action='store_true')
 
     for item in ['a', 'aaaa', 'caa', 'cname', 'mx', 'ns', 'ptr', 'srv', 'txt']:
         with self.argument_context('network dns record-set {} add-record'.format(item)) as c:
@@ -473,28 +474,30 @@ def load_arguments(self, _):
         c.argument('frontend_ip_name', help='The name of the frontend IP configuration.', completer=get_lb_subresource_completion_list('frontend_ip_configurations'))
         c.argument('floating_ip', help='Enable floating IP.', arg_type=get_enum_type(['true', 'false']))
         c.argument('idle_timeout', help='Idle timeout in minutes.')
-        c.argument('protocol', help='', arg_type=get_enum_type(TransportProtocol))
+        c.argument('protocol', help='Network transport protocol.', arg_type=get_enum_type(TransportProtocol))
         for item in ['backend_pool_name', 'backend_address_pool_name']:
             c.argument(item, options_list=('--backend-pool-name',), help='The name of the backend address pool.', completer=get_lb_subresource_completion_list('backend_address_pools'))
 
     with self.argument_context('network lb create') as c:
         c.argument('frontend_ip_zone', zone_type, min_api='2017-06-01', options_list=('--frontend-ip-zone'), help='used to create internal facing Load balancer')
-        c.argument('public_ip_zone', zone_type, min_api='2017-06-01', options_list=('--public-ip-zone'), help='used to created a new public ip for the load balancer, a.k.a public facing Load balancer')
         c.argument('validate', help='Generate and validate the ARM template without creating any resources.', action='store_true')
-        c.argument('public_ip_address_allocation', arg_type=get_enum_type(IPAllocationMethod))
-        c.argument('public_ip_dns_name', help='Globally unique DNS name for a new public IP.')
         c.argument('sku', min_api='2017-08-01', help='Load balancer SKU', arg_type=get_enum_type(LoadBalancerSkuName, default='basic'))
 
+    with self.argument_context('network lb create', arg_group='Public IP') as c:
         public_ip_help = get_folded_parameter_help_string('public IP address', allow_none=True, allow_new=True)
         c.argument('public_ip_address', help=public_ip_help, completer=get_resource_name_completion_list('Microsoft.Network/publicIPAddresses'))
-        c.ignore('public_ip_type')
+        c.argument('public_ip_address_allocation', help='IP allocation method.', arg_type=get_enum_type(IPAllocationMethod))
+        c.argument('public_ip_dns_name', help='Globally unique DNS name for a new public IP.')
+        c.argument('public_ip_zone', zone_type, min_api='2017-06-01', options_list=('--public-ip-zone'), help='used to created a new public ip for the load balancer, a.k.a public facing Load balancer')
+        c.ignore('public_ip_address_type')
 
+    with self.argument_context('network lb create', arg_group='Subnet') as c:
         subnet_help = get_folded_parameter_help_string('subnet', other_required_option='--vnet-name', allow_new=True, allow_none=True, default_none=True)
         c.argument('subnet', help=subnet_help, completer=subnet_completion_list)
         c.argument('subnet_address_prefix', help='The CIDR address prefix to use when creating a new subnet.')
-        c.argument('vnet_name', virtual_network_name_type)
+        c.argument('virtual_network_name', virtual_network_name_type)
         c.argument('vnet_address_prefix', help='The CIDR address prefix to use when creating a new VNet.')
-        c.ignore('vnet_type')
+        c.ignore('vnet_type', 'subnet_type')
 
     with self.argument_context('network lb frontend-ip') as c:
         c.argument('zone', zone_type, min_api='2017-06-01')
@@ -516,6 +519,7 @@ def load_arguments(self, _):
 
     with self.argument_context('network lb rule') as c:
         c.argument('load_distribution', help='Affinity rule settings.', arg_type=get_enum_type(LoadDistribution))
+        c.argument('probe_name', help='Name of an existing probe to associate with this rule.')
 
     # endregion
 
@@ -706,6 +710,7 @@ def load_arguments(self, _):
         c.argument('filters', type=get_json_object)
 
     with self.argument_context('network watcher flow-log') as c:
+        c.argument('nsg', help='Name or ID of the network security group.')
         c.argument('enabled', arg_type=get_three_state_flag())
 
     for item in ['list', 'stop', 'delete', 'show', 'show-status']:
@@ -725,14 +730,8 @@ def load_arguments(self, _):
         c.argument('dest_ip', help='Destination IPv4 address.')
 
     with self.argument_context('network watcher troubleshooting') as c:
-        c.argument('resource_type', options_list=['--resource-type', '-t'], id_part='resource_type', arg_type=get_enum_type(['vnetGateway', 'vpnConnection']))
-
-    with self.argument_context('network watcher troubleshooting start') as c:
         c.argument('resource', help='Name or ID of the resource to troubleshoot.')
-
-    with self.argument_context('network watcher troubleshooting stop') as c:
-        c.argument('resource', help='Name or ID of the resource to troubleshoot.')
-
+        c.argument('resource_type', help='The resource type', options_list=['--resource-type', '-t'], id_part='resource_type', arg_type=get_enum_type(['vnetGateway', 'vpnConnection']))
     # endregion
 
     # region PublicIPAddresses
@@ -778,8 +777,8 @@ def load_arguments(self, _):
 
     # region RouteTables
     with self.argument_context('network route-table') as c:
-        c.argument('route_table_name', name_arg_type, completer=get_resource_name_completion_list('Microsoft.Network/routeTables'), id_part='name')
-        c.argument('disable_bgp_propagation', arg_type=get_three_state_flag(), min_api='2017-10-01', help='Disable routes learned by BGP.')
+        c.argument('route_table_name', name_arg_type, help='Name of the route table.', completer=get_resource_name_completion_list('Microsoft.Network/routeTables'), id_part='name')
+        c.argument('disable_bgp_route_propagation', arg_type=get_three_state_flag(), min_api='2017-10-01', help='Disable routes learned by BGP.')
 
     with self.argument_context('network route-table create') as c:
         c.extra('tags')
@@ -805,7 +804,7 @@ def load_arguments(self, _):
         c.argument('monitor_path', help='Path to monitor.')
         c.argument('monitor_port', help='Port to monitor.', type=int)
         c.argument('monitor_protocol', monitor_protocol_type)
-        c.argument('profile_status', options_list=('--status',), help='Status of the Traffic Manager profile.', arg_type=get_enum_type(['Enabled', 'Disabled']))
+        c.argument('profile_status', options_list=['--status'], help='Status of the Traffic Manager profile.', arg_type=get_enum_type(ProfileStatus))
         c.argument('routing_method', help='Routing method.', arg_type=get_enum_type(['Performance', 'Weighted', 'Priority', 'Geographic']))
         c.argument('unique_dns_name', help="Relative DNS name for the traffic manager profile. Resulting FQDN will be `<unique-dns-name>.trafficmanager.net` and must be globally unique.")
         c.argument('ttl', help='DNS config time-to-live in seconds.', type=int)
@@ -830,7 +829,7 @@ def load_arguments(self, _):
         c.argument('target', help='Fully-qualified DNS name of the endpoint.')
         c.argument('target_resource_id', help="The Azure Resource URI of the endpoint. Not applicable for endpoints of type 'ExternalEndpoints'.")
         c.argument('weight', help="Weight of the endpoint when using the 'Weighted' traffic routing method. Values range from 1 to 1000.", type=int)
-        c.argument('geo_mapping', nargs='+')
+        c.argument('geo_mapping', help="Space-separated list of country/region codes mapped to this endpoint when using the 'Geographic' routing method.", nargs='+')
 
     with self.argument_context('network traffic-manager endpoint create') as c:
         c.argument('target', help='Fully-qualified DNS name of the endpoint.')
@@ -917,8 +916,8 @@ def load_arguments(self, _):
         c.argument('thumbprint', help='Certificate thumbprint.')
 
     with self.argument_context('network vnet-gateway vpn-client') as c:
-        c.argument('processor_architecture', arg_type=get_enum_type(ProcessorArchitecture))
-        c.argument('authentication_method', arg_type=get_enum_type(AuthenticationMethod))
+        c.argument('processor_architecture', help='Processor architecture of the target system.', arg_type=get_enum_type(ProcessorArchitecture))
+        c.argument('authentication_method', help='Method used to authenticate with the generated client.', arg_type=get_enum_type(AuthenticationMethod))
         c.argument('radius_server_auth_certificate', help='Public certificate data for the Radius server auth certificate in Base-64 format. Required only if external Radius auth has been configured with EAPTLS auth.')
         c.argument('client_root_certificates', nargs='+', help='Space-separated list of client root certificate public certificate data in Base-64 format. Optional for external Radius-based auth with EAPTLS')
         c.argument('use_legacy', min_api='2017-06-01', help='Generate VPN client package using legacy implementation.', arg_type=get_three_state_flag())
@@ -930,6 +929,7 @@ def load_arguments(self, _):
         c.argument('virtual_network_gateway_connection_name', options_list=('--name', '-n'), metavar='NAME', id_part='name', help='Connection name.')
         c.argument('shared_key', help='Shared IPSec key.')
         c.argument('connection_name', help='Connection name.')
+        c.argument('routing_weight', type=int, help='Connection routing weight')
         c.argument('use_policy_based_traffic_selectors', min_api='2017-03-01', help='Enable policy-based traffic selectors.', arg_type=get_three_state_flag())
 
     with self.argument_context('network vpn-connection create') as c:
@@ -939,7 +939,6 @@ def load_arguments(self, _):
             c.argument(item, arg_group='Destination')
 
     with self.argument_context('network vpn-connection update') as c:
-        c.argument('routing_weight', type=int, help='Connection routing weight')
         c.argument('enable_bgp', help='Enable BGP (Border Gateway Protocol)', arg_type=get_enum_type(['true', 'false']))
 
     with self.argument_context('network vpn-connection shared-key') as c: