+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -9,6 +9,8 @@ Release History
 * `application-gateway http-settings create/update`: Add convenience argument `--auth-certs` to attach authentication certificates. [#4910](https://github.com/Azure/azure-cli/issues/4910)
 * `ddos-protection`: Added new commands to create DDoS protection plans .
 * `vnet create/update`: Added support for `--ddos-protection-plan` to associate a VNet to a DDoS protection plan.
+* `network route-table create/update`: Fix issue with `--disable-bgp-route-propagation` flag.
+* `network lb create/update`: Removed dummy arguments `--public-ip-address-type` and `--subnet-type`.
 * `sdist` is now compatible with wheel 0.31.0
 
 2.0.27